Make tab handling configurable to support code that uses tabs (closes #1).
When you call `require("recast").run`, you can now optionally pass in a hash of options as a second argument.

The currently recognized options are `tabWidth`, an integer specifying the number of spaces per tab, and `useTabs`, a boolean specifying whether to use tabs when reprinting changed code. The defaults are `4` and `false`, respectively.

Note that lines whose indentation has not changed will be reprinted with their original spacing, no matter how crazy it was.

Amazingly, this gives a 20% speedup on the full test suite, probably from the caching of `emptyLines`.