Add a filter-branch script for normalizing authorship.
This is destructive to the git history, but the consequences are limited, because most usage of recast so far has been via NPM, and I don't have any active contributors yet.

The branch before-normalizing-authorship will preserve the state of history just before the rewrite, so you can rebase off of that and then cherry-pick commits over to the new master, if need be.