Bump NPM version to 0.2.0 in preparation for publishing.
This is a big bump because the require("recast").builder interface was completely reimplemented, now with actual type checks, and there may be subtle differences in behavior.

One difference to note: Expressions will no longer be tolerated where a Statement is expected. Make sure to wrap such Expressions with builder.ExpressionStatement to keep the type checker happy.