Always print exactly one newline after trailing comments.
The goal of preserving the whitespace that intervenes between leading
comments and the nodes that they precede was to respect whitespace
conventions in the existing code as much as possible when we fall back to
pretty-printing comments.

When we transplant a trailing in front of the node it trails, we're
already giving up on perfect preservation, so it doesn't make sense to
preserve the whitespace that originally came after the comment.