Work around Esprima SourceLocation bug that led to misplaced comments.
When a statement does not end with a semicolon, Esprima sometimes records
the wrong end location (typically the start location of the next node).

I cope with that mistake using the LocationFixer visitor in lib/parser.js,
but it only knows how to skip over whitespace, so it can't fix the
location if there's an intervening comment (as in this test).

The changes to lib/comments.js in this commit don't fix the location, but
they do prevent the comment from getting attached as a trailing comment to
the previous node.

I really should submit a patch to Esprima to fix location tracking, but
that's going to take more work.