Ignore options.sourceFileName when printing.
The root problem was that passing options.sourceFileName to the Printer
would cause any new AST nodes (i.e. nodes that shouldn't have any source
mapping) to appear to be taken from the so-named source file. If that
confusion happened before the real sourceContent was set, the new AST node
(typically the first node in the AST) would "win," providing the wrong
content for that source file.

Closes #70.