Fix printing of var decls in loop initializers
See facebook/regenerator#103.

c0320a4 broke this; variable declarators should not always be followed by semicolons (particularly in a for/in loop, where "for (var key; in obj)" is wrong). I also made sure that empty statements are printed correctly in the context of a for loop.

Test Plan:
npm test