Add a failing test showing a regression in printing statement sequences.
This regression was introduced by 554c01dc2c448e5a2b3b330c13a04356ad198321 which was intended to fix #39. My guess is that the new implementation is trying to use the original location of a node when it is not appropriate. Recast should be passing a flag or implementation function that will cause the printer to ignore original locations when printing generically.