Force reprint on `null`ing the original
This one adds the ability to force a reprint of a node, when the `original` is
set to `null`.

One of such use cases is a trailing comma in objects and array initializers.
The information about it is not written on AST (rather, it belongs to CST), so
we need just to rewrite the node and the trailing comma will gone.

```
[1, 2,]; -> [1, 2];
({x, y,}); -> ({x, y,});
```