Fix comment attachment for concise methods.
The FunctionExpression used for the .value of concise method properties is
not really a complete, syntactically valid FunctionExpression, because
(among other problems) the function keyword is missing. Esprima abuses the
FunctionExpression node in this case as a convenient means of storing the
.params and .body of the method, while ignoring its .id.

This has caused us problems before: #86, b084c6580a, 7e5bbd7925.

Previously we solved such problems by nulling out the .loc property of
degenerate/untrustworthy FunctionExpression nodes, and so we can use that
indicator again to determine whether getSortedChildNodes should consider a
given node to be an acceptable child node, or unpack it further into its
constituent child nodes.

That nullification of .loc properties occurs during the TreeCopier
traversal now, so in order for attachComments and getSortedChildNodes to
benefit from the nullification, it was necessary to move the call to
attachComments after the copy.

Fixes #126.
Fixes #127.