Allow comments to be attached to Program nodes.
Previously, a comment at the very beginning of a file would be attached to
the first statement in the body of the Program, rather than to the Program
node itself, which was a violation of the invariant that comments should
fall entirely within all ancestors of the node to which they are attached.

The test case introduced by #149 failed because the block comment was
attached to ast.program.body[0], even though the comment appeared before
ast.program.loc.start. When the printer reprinted the Program node
generically (because the length of its .body changed), the comment
appeared in the resulting replacement text, even though it also appeared
in the text surrounding the Program node.

With this change, when we reprint the Program node generically, we avoid
printing its .comments, and instead preserve the surrounding comments that
appear within the File node but outside the Program node, so the block
comment appears only once, as desired.

The balance of responsibility for printing comments is kind of subtle, but
it currently boils down to this: generically reprinted nodes lose their
comments, because the surrounding text tends to preserve those comments
anyway. Another approach would be to reprint changed nodes with their
comments, but also expand their .original.loc properties to include the
original comments (a la getTrueLoc in printer.js), so that the original
comments would get replaced by the reprinted comments.

I don't often rewrite comments during AST traversals, but this alternative
solution might handle that use case better. Food for thought.