Bump NPM version to 0.10.0 in preparation for publishing.
The minor version bump is due to changes in the representation of comment
objects, which now have .leading and .trailing boolean properties to
indicate their relationship to the node. The .comments property is now
just an array of these comment objects, whereas previously it was an
object with .leading, .trailing, and .dangling properties, which were
separate arrays that contained comment objects. The .comments property is
also no longer non-enumerable, so there's a chance that third-party
traversal code will have to begin handling (or ignoring) that property.

Closes #153.