Attach initial comments to program.body[0] instead of program.
To preserve the invariant that comments should fall entirely within all
ancestors of the node to which they are attached, it was necessary to
expand program.loc to include all comments in the program.

This reverts 053630a744253949ce2609be9e158f76065e07ce, now that we can
solve the same problem in a better way.

Closes #159.