Allow Comment nodes to appear in place of non-Comment nodes.
This commit just makes it possible to pretend that a Comment can appear as
a Statement or an Expression. Notice that the test has to move comments
around manually to get them to be printed in the right places. The next
step is to make the comment attachment code actually insert dangling
comments in reasonable places, rather than marking them as neither
.leading nor .trailing, which has the unfortunate consequence of rendering
them un-pretty-printable.

Inspired by https://github.com/benjamn/recast/issues/189