Use esprima rather than esprima-fb as the default parser.
Tests that relied on esprima-fb-specific behavior continue to use
esprima-fb, but recast.parse by default will now use esprima. Note that
there are other tests (test/babylon.js) that use the babylon parser
instead of esprima(-fb), so the strategy of using different parsers for
different tests already has precedent.

The more general point to take from this change is that recast is tolerant
of many different parsers, since the reprinting algorithm is not tied to
any particular AST format, and the pretty-printer supports a wide variety
of AST types, not just the types that any single parser might produce.

Closes #178.