Avoid an ASI bug by sometimes wrapping return expressions in parens
Fixes #362

If the argument to a return statement starts with a comment-only line, then we
now wrap it in parens to avoid the `return` getting parsed as its own line:

```js
function f() {
  return (
    // Foo
    3
  );
}
```

This problem can be introduced by adding a leading comment to any leftmost child
of the return argument, so we need to be more defensive about when we reformat
`return` statements. For now, we reprint the `ReturnStatement` node (but leave
unchanged children intact) when any child had changes.