Tolerate duplicates when scanning new/old ASTs for reprints.
Duplicate nodes are acceptable in an AST (though it should technically be
called an ASD, since duplication makes it a Directed Acyclic Graph or
DAG), as long as the node is not its own ancestor. That's why we only
search the current FastPath stacks for duplicates, instead of keeping
track of every node we've ever seen.

If the node is its own ancestor, that's a logical error, and we cannot
print the AST because it's neither a tree nor acyclic.

Note that 0c2024064b50f7181299df44c15b566f3fd21af9 helps avoid this
problem for "private" properties added to the AST by tools like Babel,
since those properties typically begin with an '_' character.

Related: #374