- bug (fixed): end-of-comments in array get **human-visually** attached to wrong item - bug (fixed): wrongly injects semicolon/comma/... in trialing comment
## HOTFIX 2: Part One / Part Two / Part Three

Mark up the start and end of LINE-comments,
so the caller can detect one dangling and thus obscuring
a comma, semicolon or other character being appended
at the end of the CODE LINE then.

...

The Third Part a.k.a. The End

Now we hunt for all ETX markers, check if they are
followed by a terminator character on the same line
and if they are, we patch that character into
the previous STX character slot.

Meanwhile we clear out all ETX character slots
with spaces, which should keep any accompanying
sourcemap nicely intact! :-)