Add better support for importKind and exportKind
In addition to an importKind on an ImportDeclaration
(`import type {Foo} from 'bar'`), it's also valid to have an importKind
on an ImportSpecifier (`import {type Foo} from 'bar'`).

Additionally, it's valid to have an exportKind on an
ExportNamedDeclaration (`export type {Foo};`).

This adds support for both these cases, and a couple simple tests.