Support options.avoidRootParens in main print helper function.
Normally the genericPrint function will attempt to wrap its result with
parentheses if path.needsParens(), but that's not what we want if we're
printing the path in order to patch it into a location that already has
parentheses, so it's important to be able to disable this behavior in such
special cases.

Note that children printed recursively by the genericPrint[NoParens]
functions will still be wrapped with parentheses if necessary, since this
option applies to the printing of the root node only.

Should help with #327.