Drastically simplify FastPath#hasParens.
The complicated logic that I implemented previously was an approximation
of a more fundamental and ultimately much simpler decision problem: some
nodes need parentheses only because they can't come first in the enclosing
statement, due to parsing ambiguities (e.g. FunctionExpression and
ObjectExpression nodes). However, these nodes do not need to be followed
immediately by a closing parenthesis, because the presence of an opening
parenthesis is enough to resolve the parsing ambiguity.