Straighten out handling of Flow function types.
This way is simpler to reason about: it's fewer conditionals, and
less peeking up and down the tree to try to anticipate what other
nodes will do.  And it works the same on all existing tests.

It also works correctly on a number of other test cases where the
existing logic breaks.  Add a bunch of those, and a few others.