test: Make Flow internal-slots test realistic, by round-tripping
I started by converting this test to a round-trip test with the same
source code verbatim, just intending to make the test simpler and
easier to read.

But then it turned out that that version failed, because this code
doesn't parse as this tree!  It parses as something wacky, which spits
out wacky output.  And in fact Flow itself doesn't accept this code;
it gives a series of parse errors.

Instead, write a version of the code that Flow accepts (at least as
far as parsing it -- it gives a later-stage error saying it doesn't
recognize these particular internal slots), mainly by saying
`declare class C` instead of `type A =`.

So this version is not only simpler but more correct.

Also add a few more cases that weren't covered in the previous test.
Now that it only takes one line of code per case, it's easy to be
thorough.