flow: Fix missing parens on parameter for some functions
In each of these test cases, the parens around the function's one
parameter's type were getting dropped, and that changed the meaning.

Comparing this logic with what's in Prettier, the key piece there --
in src/language-js/print/function-parameters.js , at the definition
of isFlowShorthandWithOneArg in printFunctionParameters -- is that
the parens will only be dropped if the type of the one parameter is
"simple".

That, in turn, basically means that its top-level node comes from a
particular list of node types, like keywords and type references.

As a minimal fix, require the type to be "simple" at least in a
weaker sense: it's not a function, union, or intersection type.
In each of those, the parens are essential: without them, the type
can end up parsing differently.  We still drop parens in some
cases where Prettier includes them, like object types.

The test case here that gets deleted, rather than uncommented, is
actually invalid Flow syntax: it's a plain `function` statement,
rather than `declare function`, so it's required to have a name for
the parameter.  One of the other cases is exactly the same but with
`declare`, and that one now works.