flow: Fix printing of getters and setters
This solution feels pretty gross, especially around
isArrowFunctionTypeAnnotation, and I feel like there should be a
cleaner way to handle it.  But I don't immediately see how... and
looking at Prettier's implementation, it's pretty similar and not
really any cleaner, so that's some validation that there isn't an
obvious better way.

(Specifically, Pretter has a very similar set of checks on the parent,
checking if its type is one of ObjectTypeCallProperty or
ObjectTypeProperty or so on; but instead of this particular line where
we look at `parent.kind`, it checks whether the node and its parent
start at the same location.  That feels like it's using some awfully
coincidental facts about the concrete syntax of things like methods
and getters and setters.)