deps: Bump ast-types version to 0.15.3-0.tsflower.5
This fixes that issue with `kind` missing on ObjectTypeProperty.

Then delete the ts-expect-error comment it made necessary... and fix
the other type error that was hiding behind that error, oops.  (Even
when the ts-expect-error wasn't present.  Gee, thanks, TS.)  The old
code was harmless, but would just check `kind` on internal slots,
which don't have that property.