parens: Greatly simplify hasParens
I don't quite see why the justNeedsOpeningParen logic would be
necessary.  If having just an opening parenthesis is good enough for
those nodes (and indeed it seems like it is), then it seems like it
should be good enough for any node.  I.e., we could treat
justNeedsOpeningParen as always true.

And if we do so, all the tests still pass.

So I'm now pretty confident that we can in fact cut that logic out.
This code gets quite a bit simpler when we do, so that's nice!

Then in particular, cutting out this pair of calls to firstInStatement
and canBeFirstInStatement leaves just one remaining call site for each
of them.  That makes things a lot simpler for us when we go to make
changes to how those functions work -- which we're about to do, in
order to fix some bugs.