parens [nfc]: Cut impossible BlockStatement case from firstInStatement
To get here, the child would have to already be a statement,
so we'd have stopped at the previous step.

All tests still pass.  In fact, all tests still pass if instead
of deleting this one replaces the body with `assert.fail()`.
So that adds empirical evidence that this case is impossible.