parens: Fix on `let [`; nail down first-in-statement logic to spec
This is a fun edge case I discovered from reading the ES spec,
in particular here:
  https://tc39.es/ecma262/#prod-ExpressionStatement

while working out whether there were cases we were missing in this
logic, beyond the three we had.

I believe this turns out to be the only such case.  Add comments
explaining why, with references to the language spec.